requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
requests>=2.31.0
pyaudio>=0.2.13  # for microphone input via SpeechRecognition
